Test (#6)
* Fix Login form loading

* Front end build logic

* Front end build

* Test pipeline

* Add lints and test

* Test env vars

* Change secrets to env

* Test env var exposing

* Test again

* Build docker / mongo

* Enable lints and tests

* Try to deploy

* Add production cors origins

* Disable health check

* Update Procfile

* Test deployment

* Update start-prod command

* New try

* Try once more

* Once more

* Revert Procfile

* Test container deployment

* Change image name

* Add tag to image

* Try this

* Test

* Test aha

* Try again

* test

* Try once more

* Fix

* TEST

* TESTTER

* Try one package.json

* New yarn.lock file

* Update port

* Test worker

* Try docker-compose in release

* Try once more

* Try this

* Try

* Add Procfile

* Just

* Update

* Update Dockerfile

* Test deployment

* Update again

* I believe

* Vite build

* This time it works

* Change vite command to yarn

* Test

* Fix env var

* Test2

* New env vars for heroku deploy

* Tadaa

* TEST AGAIN

* TEST2

* Test3

* Huoh

* TESTER

* TEST

* Quick test

* Tadaa

* Kukkuu

* Test again

* Another

* All yarn installs

* Now

* Fiba perskules

* FGDSFDS

* Duup

* TADAA32

* Kukkuluuruu

* Try once more

* FIX

* TEst these

* Update

* Update vite server

* Ful deployment

* Fix deployment script

* Try again

* Try to fail

* Try once more

* Fail tests on purpose

* Test

* Fail front test

* Once again

* Tester

* Testing one two three

* Fix tests

* Fix for frontend jest

+++ shared/formData/adminSettingsFormData.js
@@ -84,7 +84,7 @@ const adminSettingsFormData = {
             descriptionId: 'session_age_desc',
             defaultValue: 60, // one hour
             required: true,
-            settingType: 'integer',
+            settingType: 'float',
             regex: '[0-9]+$',
           },
           {
@@ -95,7 +95,7 @@ const adminSettingsFormData = {
             descriptionId: 'remember_me_session_age_desc',
             defaultValue: 10080, // 7 days
             required: true,
-            settingType: 'integer',
+            settingType: 'float',
             regex: '[0-9]+$',
           },
         ],
